requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.26.2
pytest==7.4.3
pytest-flask==1.3.0
Werkzeug==3.0.1
